feat: allow "*" as a wildcard host in allowlist matcher
Short-circuit MatchGlob when the configured pattern is "*" so a single
"*" entry allows any host. Also moves TestMatchGlob out of the dns
package and into hostmatch where it belongs.
